# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Past Paper Analyzer API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /syllabus:
    get:
      operationId: listSyllabi
      tags: [syllabus]
      summary: List all syllabi
      responses:
        "200":
          description: List of syllabi
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Syllabus"
    post:
      operationId: uploadSyllabus
      tags: [syllabus]
      summary: Upload a syllabus
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/UploadSyllabusBody"
      responses:
        "201":
          description: Syllabus uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Syllabus"
  /syllabus/{id}:
    get:
      operationId: getSyllabus
      tags: [syllabus]
      summary: Get a syllabus by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Syllabus details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Syllabus"
...
  /topics:
    get:
      operationId: listTopics
      tags: [topics]
      summary: List all topics with scores
      parameters:
        - in: query
          name: subject
          schema:
            type: string
        - in: query
          name: analysisId
          schema:
            type: integer
      responses:
        "200":
          description: List of topics
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Topic"
  /topics/high-yield:
    get:
      operationId: getHighYieldTopics
      tags: [topics]
      summary: Get high-yield topics ranked by importance
      parameters:
        - in: query
          name: subject
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: High-yield topics leaderboard
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Topic"
...
components:
  schemas:
...
    Syllabus:
      type: object
      properties:
        id:
          type: integer
        subject:
          type: string
        fileName:
          type: string
        topics:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
      required: [id, subject, fileName, topics, createdAt]
    UploadSyllabusBody:
      type: object
      properties:
        subject:
          type: string
        file:
          type: string
          format: binary
      required: [subject, file]
...
    Topic:
      type: object
      properties:
        id:
          type: integer
        analysisId:
          type: integer
        name:
          type: string
        subject:
          type: string
        frequency:
          type: integer
        frequencyScore:
          type: number
        recencyScore:
          type: number
        marksWeightage:
          type: number
        importanceScore:
          type: number
        questionTypes:
          type: array
          items:
            type: string
        yearsAppeared:
          type: array
          items:
            type: integer
        difficultyLevel:
          type: string
          enum: [easy, medium, hard]
        inSyllabus:
          type: boolean
      required: [id, analysisId, name, subject, frequency, importanceScore, difficultyLevel]
...